+++ apps/lockfile-explorer/src/utils/init.ts
