+++ apps/lockfile-explorer/src/utils/shrinkwrap.ts
